Stop SSE connections when the JWT token expires (#24)
With this PR, the SSE connection is closed when the JWT token used to
authenticate actually expires. Previously, the JWT token expiration was
only checked when the SSE connection is established.

Benefits:
- Improve overall security by preventing over-usage of JWT tokens,
- Allow backends that issue tokens to have their own expiration policy
(which must still be lower than the global max lifetime configured in
Neurow). For exemple, it allows the backend to use an `exp` that matches
the duration when users are actually considered as active on the
frontend.