Various observability fixes (#40)
- `cowboy_telemetry` both provides HTTP status like "200", "200 OK", or
"403 forbidden", and they are converted to prometheus labels as `200`,
`200_ok`, `403_forbidden`... In the PR, only the integer part is kept to
only export labels like `200`, `403` ...
- Excludes the health check, the metric export, and the favicon from the
set of HTTP resources that are monitored. The goal is to have a better
understanding of the behavior of HTTP resources related to message
publishing.